+++ Engine/RHI/Include/RHI/PipelineLayout.h
@@ -5,15 +5,14 @@
 #ifndef EXPLOSION_RHI_PIPELINE_LAYOUT_H
 #define EXPLOSION_RHI_PIPELINE_LAYOUT_H
 
-#include <cstddef>
-
+#include <cstdint>
 #include <Common/Utility.h>
 
 namespace RHI {
     class BindGroupLayout;
 
     struct PipelineLayoutCreateInfo {
-        size_t bindGroupNum;
+        uint32_t bindGroupNum;
         const BindGroupLayout* bindGroupLayouts;
     };
